requirements.txt
openai-whisper==20231117
textdistance==4.6.2
gradio==4.44.0
torch==2.1.0
torchaudio==2.1.0
